+++ NEWS.md
@@ -1,5 +1,9 @@
 # exclude (development version)
 
+* Refactored internal code so that common code is extracted into the update method of exclude class.
+
+* Added examples to all functions.
+
 # exclude 0.0.1
 
 * Fixed examples in function documentation so that the environment is kept clean.